# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/usuarios": {
      "get": {
        "operationId": "UsuariosController_listar",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Cantidad de elementos por página (máx. 100)"
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Número de página (empieza en 1)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de usuarios"
          },
          "400": {
            "description": "Query inválida"
          }
        },
        "summary": "Listar usuarios paginados",
        "tags": [
          "usuarios"
        ]
      },
      "post": {
        "operationId": "UsuariosController_crear",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nombre",
                  "email",
                  "password"
                ],
                "properties": {
                  "nombre": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 255
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 255
                  },
                  "activo": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado"
          },
          "400": {
            "description": "Campos inválidos"
          },
          "409": {
            "description": "Email duplicado"
          }
        },
        "summary": "Crear un usuario",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/usuarios/{id}": {
      "get": {
        "operationId": "UsuariosController_obtenerPorId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado"
          },
          "400": {
            "description": "Id inválido"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        },
        "summary": "Obtener un usuario por id",
        "tags": [
          "usuarios"
        ]
      },
      "patch": {
        "operationId": "UsuariosController_actualizar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nombre",
                  "email",
                  "password"
                ],
                "properties": {
                  "nombre": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 255
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 255
                  },
                  "activo": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado"
          },
          "400": {
            "description": "Campos inválidos"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "409": {
            "description": "Email duplicado"
          }
        },
        "summary": "Actualizar un usuario",
        "tags": [
          "usuarios"
        ]
      },
      "delete": {
        "operationId": "UsuariosController_eliminar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado"
          },
          "400": {
            "description": "Id inválido"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        },
        "summary": "Eliminar un usuario",
        "tags": [
          "usuarios"
        ]
      }
    }
  },
  "info": {
    "title": "API de Tarea",
    "description": "API REST + tRPC con contratos Zod y Drizzle ORM",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {}
  }
}